requirements.txt
# Monitor de EPI — câmeras IP / RTSP + visão computacional
opencv-python-headless>=4.8.0
ultralytics>=8.1.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0
httpx>=0.27.0
